vm: refactor TxContext to an interface
